requirements.txt
gradio
langchain
openai
tiktoken
langchain-openai
langchain-google-genai
langchain-anthropic
langchain-core
pyyaml
pydantic
networkx
pyvis